requirements.txt
fastapi
uvicorn
paramiko
python-dotenv
requests
